Keywords in notifications #94
- Fixed tests
- Added notify_keyword test
- refactored notify_keywords and notify_favorites to extend bas/notify
- updated mode marked_keyword_title and marked_keyword_description to accept empty title
