+++ app/templates/mails/tender_update.html
@@ -1,6 +1,6 @@
 {% extends 'mails/layout.html' %}
 
-{% block header %}Favorite tender(s) update{% endblock %}
+{% block header %}{{ notification_type }} tender(s) update{% endblock %}
 
 {% block content %}
   <ol class="tender-list">
